VIX-3232 Move Tagged Filter Setup dialog to base class
VIX-3232 Move Tagged Filter Setup dialog to base class
